Implement search functionality and add styling
This tweaks the home view to add a search form and implements searching
on the images index page by search_type and a sample query. Users can
now search by matching a substring for title, caption, and autogenerated
description or enter a space separated string of words to specify tags.
The table description row for tags was also altered to provide quick
links to similar tag searches. A conditional check was added in the case
of no results to display an info message.